fix: move serialization to a plugin
