+++ src/compiler/plugins/references.plugin.ts
@@ -25,3 +25,25 @@ export class ExtractReferencesPlugin implements CompilerPlugin<ContextWithRefere
     };
   }
 }
+
+export class ReplacePlaceholdersPlugin implements CompilerPlugin<ContextWithReferences, ContextWithReferences> {
+  run(context: ContextWithReferences): ContextWithReferences {
+    const { bytes, references } = context;
+    const output = [];
+
+    for (let index = 0; index < bytes.length; ) {
+      const node = bytes[index];
+
+      if (isPlaceholder(node)) {
+        output.push(...numberToInt32(references.get(node.name)));
+        index += 4;
+        continue;
+      }
+
+      output.push(node);
+      index++;
+    }
+
+    return { ...context, bytes: output };
+  }
+}
